refactor Github event handlers into helper module - handle_github_issue_opened - handle_github_issue_edited - handle_github_issue_closed
